+++ antarest/study/storage/rawstudy/model/filesystem/root/input/areas/areas.py
@@ -11,10 +11,10 @@
 # This file is part of the Antares project.
 from typing_extensions import override
 
-from antarest.matrixstore.matrix_uri_mapper import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.folder_node import FolderNode
 from antarest.study.storage.rawstudy.model.filesystem.inode import TREE
+from antarest.study.storage.rawstudy.model.filesystem.matrix.matrix_storage_context import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.root.input.areas.item.item import InputAreasItem
 from antarest.study.storage.rawstudy.model.filesystem.root.input.areas.list import InputAreasList
 from antarest.study.storage.rawstudy.model.filesystem.root.input.areas.sets import InputAreasSets
