+++ antarest/study/storage/rawstudy/model/filesystem/root/input/areas/list.py
@@ -14,9 +14,9 @@
 from typing_extensions import override
 
 from antarest.core.utils.archives import extract_lines_from_archive
-from antarest.matrixstore.matrix_uri_mapper import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.inode import INode
+from antarest.study.storage.rawstudy.model.filesystem.matrix.matrix_storage_context import MatrixStorageContext
 
 AREAS_LIST_RELATIVE_PATH = "input/areas/list.txt"
