+++ antarest/study/storage/rawstudy/model/filesystem/root/input/renewables/clusters.py
@@ -14,11 +14,11 @@ from typing_extensions import override
 from antarest.core.serde.ini_common import any_section_option_matcher
 from antarest.core.serde.ini_reader import LOWER_CASE_PARSER, IniReader
 from antarest.core.serde.ini_writer import LOWER_CASE_SERIALIZER, IniWriter
-from antarest.matrixstore.matrix_uri_mapper import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.folder_node import FolderNode
 from antarest.study.storage.rawstudy.model.filesystem.ini_file_node import IniFileNode
 from antarest.study.storage.rawstudy.model.filesystem.inode import TREE
+from antarest.study.storage.rawstudy.model.filesystem.matrix.matrix_storage_context import MatrixStorageContext
 
 _VALUE_PARSERS = {any_section_option_matcher("group"): LOWER_CASE_PARSER}
 _VALUE_SERIALIZERS = {any_section_option_matcher("group"): LOWER_CASE_SERIALIZER}
