+++ antarest/study/storage/rawstudy/model/filesystem/root/output/simulation/mode/common/links.py
@@ -13,10 +13,10 @@
 
 from typing_extensions import override
 
-from antarest.matrixstore.matrix_uri_mapper import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.folder_node import FolderNode
 from antarest.study.storage.rawstudy.model.filesystem.inode import TREE
+from antarest.study.storage.rawstudy.model.filesystem.matrix.matrix_storage_context import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.root.output.simulation.mode.common.link import (
     OutputSimulationLinkItem,
 )
