+++ antarest/study/storage/rawstudy/model/filesystem/root/user/expansion/settings.py
@@ -12,9 +12,9 @@
 
 from antarest.core.serde.ini_reader import SimpleKeyValueReader
 from antarest.core.serde.ini_writer import SimpleKeyValueWriter
-from antarest.matrixstore.matrix_uri_mapper import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.ini_file_node import IniFileNode
+from antarest.study.storage.rawstudy.model.filesystem.matrix.matrix_storage_context import MatrixStorageContext
 
 
 class ExpansionSettings(IniFileNode):
