+++ tests/storage/rawstudies/test_factory.py
@@ -13,10 +13,10 @@
 from unittest.mock import Mock
 
 from antarest.core.interfaces.cache import study_config_cache_key
-from antarest.matrixstore.matrix_uri_mapper import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.config.files import build
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfigDTO
 from antarest.study.storage.rawstudy.model.filesystem.factory import StudyFactory
+from antarest.study.storage.rawstudy.model.filesystem.matrix.matrix_storage_context import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.root.filestudytree import FileStudyTree
 from tests.storage.rawstudies.samples import ASSETS_DIR
