+++ tests/storage/repository/filesystem/root/output/simulation/mode/common/test_binding_const.py
@@ -17,9 +17,9 @@ from unittest.mock import Mock
 
 import pytest
 
-from antarest.matrixstore.matrix_uri_mapper import MatrixStorageContext
 from antarest.study.model import MatrixFrequency
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
+from antarest.study.storage.rawstudy.model.filesystem.matrix.matrix_storage_context import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.matrix.output_series_matrix import (
     BindingConstraintOutputSeriesMatrix,
 )
