+++ tests/storage/repository/filesystem/test_lazy_node.py
@@ -15,9 +15,9 @@ from unittest.mock import Mock
 
 from typing_extensions import override
 
-from antarest.matrixstore.matrix_uri_mapper import MatrixStorageContext
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.lazy_node import LazyNode
+from antarest.study.storage.rawstudy.model.filesystem.matrix.matrix_storage_context import MatrixStorageContext
 from tests.storage.repository.filesystem.matrix.test_matrix_node import MockInputSeriesMatrix
